+++ src/stories/clusters.stories.ts
@@ -1,7 +1,14 @@
 import type { Meta, StoryObj } from '@storybook/html'
-import { CosmosStoryProps, createCosmos } from '@/graph/stories/create-cosmos'
+import { CosmosStoryProps } from '@/graph/stories/create-cosmos'
+import { WithLabelsStory } from './clusters/with-labels'
+import { WormStory } from './clusters/worm'
+import { RadialStory } from './clusters/radial'
 
-import { generateMeshData } from './generate-mesh-data'
+import createCosmosRaw from './create-cosmos?raw'
+import generateMeshDataRaw from './generate-mesh-data?raw'
+import withLabelsStory from './clusters/with-labels?raw'
+import wormStory from './clusters/worm?raw'
+import radialStory from './clusters/radial?raw'
 
 const meta: Meta<CosmosStoryProps> = {
   title: 'Clusters',
@@ -10,57 +17,38 @@ const meta: Meta<CosmosStoryProps> = {
       disable: true,
     },
   },
-  render: (args) => {
-    return createCosmos(args)
-  },
+}
+
+type Story = StoryObj<CosmosStoryProps>;
+
+const sourceCodeAddonParams = [
+  { name: 'create-cosmos', code: createCosmosRaw },
+  { name: 'generate-mesh-data', code: generateMeshDataRaw },
+]
+
+export const Worm = WormStory as Story
+Worm.parameters = {
+  sourceCode: [
+    { name: 'Story', code: wormStory },
+    ...sourceCodeAddonParams,
+  ],
+}
+
+export const Radial = RadialStory as Story
+Radial.parameters = {
+  sourceCode: [
+    { name: 'Story', code: radialStory },
+    ...sourceCodeAddonParams,
+  ],
+}
+
+export const WithLabels = WithLabelsStory as Story
+WithLabels.parameters = {
+  sourceCode: [
+    { name: 'Story', code: withLabelsStory },
+    ...sourceCodeAddonParams,
+  ],
 }
 
 // eslint-disable-next-line import/no-default-export
 export default meta
-type Story = StoryObj<CosmosStoryProps>;
-
-const wormData = generateMeshData(100, 100, 1000, 1.0)
-export const worm: Story = {
-  args: {
-    simulationGravity: 0.5,
-    simulationRepulsion: 1,
-    simulationLinkSpring: 1,
-
-    pointPositions: wormData.pointPositions,
-    pointColors: wormData.pointColors,
-    pointClusters: wormData.clusters,
-    links: wormData.links,
-    linkColors: wormData.linkColors,
-  },
-}
-
-const radialData = generateMeshData(100, 100, 100, 1.0)
-export const radial: Story = {
-  args: {
-    pointPositions: radialData.pointPositions,
-    pointColors: radialData.pointColors,
-    pointSizes: radialData.pointSizes,
-    pointClusters: radialData.clusters,
-
-    links: radialData.links,
-    linkColors: radialData.linkColors,
-    linkWidths: radialData.linkWidths,
-
-    clusterPositions: radialData.clusterPositions,
-    clusterStrength: radialData.clusterStrength,
-  },
-}
-
-const withoutLinksData = generateMeshData(100, 100, 15, 1.0)
-export const withoutLinks: Story = {
-  args: {
-    simulationGravity: 0.5,
-    simulationLinkSpring: 1,
-
-    pointPositions: withoutLinksData.pointPositions,
-    pointColors: withoutLinksData.pointColors,
-    pointClusters: withoutLinksData.clusters,
-
-    showClusterLabels: true,
-  },
-}
