+++ src/stories/create-cosmos.ts
@@ -13,10 +13,9 @@ export type CosmosStoryProps = GraphConfigInterface & {
   pointClusters?: number[];
   clusterPositions?: number[];
   clusterStrength?: Float32Array;
-  showClusterLabels?: boolean;
 }
 
-export const createCosmos = (props: CosmosStoryProps): HTMLDivElement => {
+export const createCosmos = (props: CosmosStoryProps): { div: HTMLDivElement; graph: Graph} => {
   const div = document.createElement('div')
   // const canvas = document.createElement('canvas')
   div.style.height = '100vh'
@@ -47,78 +46,23 @@ export const createCosmos = (props: CosmosStoryProps): HTMLDivElement => {
     ...props,
   }
 
-  // If we initialize the graph before the canvas is added to the DOM, nothing happens
-  requestAnimationFrame(() => {
-    const graph = new Graph(div, config)
+  const graph = new Graph(div, config)
 
-    graph.setPointPositions(props.pointPositions)
-    graph.setPointColors(props.pointColors)
-    if (props.pointSizes) graph.setPointSizes(props.pointSizes)
+  graph.setPointPositions(props.pointPositions)
+  graph.setPointColors(props.pointColors)
+  if (props.pointSizes) graph.setPointSizes(props.pointSizes)
 
-    if (props.links) graph.setLinks(props.links)
-    if (props.linkColors) graph.setLinkColors(props.linkColors)
-    if (props.linkWidths) graph.setLinkWidths(props.linkWidths)
-    // if (props.linkStrength) graph.setLinkStrength(props.linkStrength)
+  if (props.links) graph.setLinks(props.links)
+  if (props.linkColors) graph.setLinkColors(props.linkColors)
+  if (props.linkWidths) graph.setLinkWidths(props.linkWidths)
+  // if (props.linkStrength) graph.setLinkStrength(props.linkStrength)
 
-    if (props.pointClusters) graph.setPointClusters(props.pointClusters)
-    if (props.clusterPositions) graph.setClusterPositions(props.clusterPositions)
-    if (props.clusterStrength) graph.setPointClusterStrength(props.clusterStrength)
+  if (props.pointClusters) graph.setPointClusters(props.pointClusters)
+  if (props.clusterPositions) graph.setClusterPositions(props.clusterPositions)
+  if (props.clusterStrength) graph.setPointClusterStrength(props.clusterStrength)
 
-    graph.zoom(0.9)
-    graph.render()
+  graph.zoom(0.9)
+  graph.render()
 
-    if (props.showClusterLabels) {
-      const clusterLabelDivs: HTMLDivElement[] = []
-      function updateClusterLabels (graph: Graph): void {
-        const clusterPositions = graph.getClusterPositions()
-        const nClusters = clusterPositions.length / 2
-        if (nClusters === 0) return
-        if (clusterLabelDivs.length === 0) {
-          for (let i = 0; i < nClusters; i++) {
-            const clusterLabelDiv = document.createElement('div')
-            const contentLabel = document.createElement('p')
-            clusterLabelDiv.appendChild(contentLabel)
-            clusterLabelDiv.style.position = 'absolute'
-            clusterLabelDiv.style.pointerEvents = 'none'
-
-            contentLabel.style.fontFamily = [
-              '"Nunito Sans"',
-              '-apple-system',
-              '".SFNSText-Regular"',
-              '"San Francisco"',
-              'BlinkMacSystemFont',
-              '"Segoe UI"',
-              '"Helvetica Neue"',
-              'Helvetica',
-              'Arial',
-              'sans-serif',
-            ].join(', ')
-            contentLabel.style.fontWeight = 'bold'
-            contentLabel.style.color = 'white'
-            contentLabel.style.transform = 'translate(-50%, -100%)'
-            contentLabel.innerText = `Cluster ${i + 1}`
-
-            div.appendChild(clusterLabelDiv)
-            clusterLabelDivs[i] = clusterLabelDiv
-          }
-        }
-
-        for (let i = 0; i < nClusters; i++) {
-          const clusterPosition = clusterPositions.slice(i * 2, i * 2 + 2)
-          const x = clusterPosition[0]
-          const y = clusterPosition[1]
-          const clusterLabelDiv = clusterLabelDivs[i] as HTMLDivElement
-          const screenXY = graph.spaceToScreenPosition([x ?? 0, y ?? 0])
-          clusterLabelDiv.style.top = `${screenXY[1]}px`
-          clusterLabelDiv.style.left = `${screenXY[0]}px`
-        }
-      }
-      graph.setConfig({
-        onZoom: updateClusterLabels.bind(this, graph),
-        onSimulationTick: updateClusterLabels.bind(this, graph),
-      })
-    }
-  })
-
-  return div
+  return { div, graph }
 }
