+++ src/stories/experiments/full-mesh.ts
@@ -3,10 +3,11 @@ import { generateMeshData } from '../generate-mesh-data'
 
 export const FullMeshStory = (): HTMLDivElement => {
   const { pointPositions, links, pointColors } = generateMeshData(40, 30, 15, 1.0)
-
-  return createCosmos({
+  const { div } = createCosmos({
     pointPositions,
     links,
     pointColors,
   })
+
+  return div
 }
