+++ src/stories/experiments/mesh-with-holes.ts
@@ -3,10 +3,11 @@ import { generateMeshData } from '../generate-mesh-data'
 
 export const MeshWithHolesStory = (): HTMLDivElement => {
   const { pointPositions, links, pointColors } = generateMeshData(40, 80, 15, 0.8)
-
-  return createCosmos({
+  const { div } = createCosmos({
     pointPositions,
     links,
     pointColors,
   })
+
+  return div
 }
